+++ examples/rust/src/bin/tx-blocktime.rs
@@ -114,7 +114,8 @@ async fn main() -> anyhow::Result<()> {
                 match msg.update_oneof {
                     Some(UpdateOneof::Transaction(tx)) => {
                         let entry = messages.entry(tx.slot).or_default();
-                        let sig = Signature::new(tx.transaction.unwrap().signature.as_slice())
+                        let sig = Signature::try_from(tx.transaction.unwrap().signature.as_slice())
+                            .expect("valid signature from transaction")
                             .to_string();
                         if let Some(timestamp) = entry.0 {
                             info!("received txn {} at {}", sig, timestamp);
@@ -125,7 +126,7 @@ async fn main() -> anyhow::Result<()> {
                     Some(UpdateOneof::BlockMeta(block)) => {
                         let entry = messages.entry(block.slot).or_default();
                         entry.0 = block.block_time.map(|obj| {
-                            DateTime::from_utc(
+                            DateTime::from_naive_utc_and_offset(
                                 NaiveDateTime::from_timestamp_opt(obj.timestamp, 0).unwrap(),
                                 Utc,
                             )