+++ yellowstone-grpc-geyser/src/lib.rs
@@ -7,5 +7,4 @@ pub mod filters;
 pub mod grpc;
 pub mod plugin;
 pub mod prom;
-pub mod proto;
 pub mod version;